Add files via upload
This is model trained on graphsage on sampled mag dataset. I split the 2w data into train, validatino and test, seperated by year of paper like 2013, 2015. The accuracy (47% for training 100 epoches) is higher than the 2005.00687v7 arxiv paper, it provided only 31.53% . Maybe my sampled dataset is better connected, so it is easier to predict the node label.
This is the first version, training is really faster than I thought, it only takes minutes, I will finish the other part later.